requirements.txt
absl-py==1.4.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.1.0
brotlipy==0.7.0
cachetools==5.3.1
certifi
cffi
charset-normalizer
click==8.1.4
cloudpickle
comm==0.2.0
contourpy==1.1.0
cryptography
cycler==0.11.0
cytoolz
dask
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.3
fonttools==4.40.0
fqdn==1.5.1
fsspec==2024.3.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
huggingface-hub==0.22.2
hydra-core==1.3.2
idna
imagecodecs
imageio
importlib-metadata==6.8.0
importlib-resources==6.0.0
iopath==0.1.9
ipykernel==6.26.0
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
kiwisolver==1.4.4
kornia==0.6.12
llvmlite==0.40.1
locket
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.6
mkl-random
mkl-service==2.4.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx
notebook==7.0.6
notebook_shim==0.2.3
numba==0.57.1
numpy
oauthlib==3.2.2
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.76
overrides==7.4.0
packaging
pandocfilters==1.5.0
parso==0.8.3
partd
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.1
pooch
portalocker==2.7.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser
Pygments==2.16.1
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets
PyYAML
pyzmq==25.1.1
qudida==0.0.4
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.2
scikit-image
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.2
sewar==0.4.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
toolz
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.13.0
transformers==4.39.3
types-python-dateutil==2.8.19.14
typing_extensions
uri-template==1.3.0
urllib3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.6
yacs==0.1.8
zipp==3.16.0
